Moved file_format to glue.utils.misc
